fix: mypy 型エラー4件修正 — delta_color Literal + beta float cast
Session-Context: CI mypy チェック失敗の修正
Changes-Made:
- style_profile.py: cov/bench_var の戻り値に float() キャスト追加 (no-any-return)
- tab_insights.py: delta_color 引数に Literal['normal','inverse'] 型注釈追加 (arg-type x3)
- tab_insights.py: typing に Literal インポート追加

+++ src/core/behavior/style_profile.py
@@ -154,7 +154,7 @@ def _compute_beta(
             return None
 
         cov = combined["port"].cov(combined["bench"])
-        return round(cov / bench_var, 2)
+        return round(float(cov / bench_var), 2)
     except Exception as exc:  # pragma: no cover
         logger.debug("_compute_beta: failed — %s", exc)
         return None
